[context - main] change theme -> app context
