+++ src/context/AppContext.jsx
@@ -1,9 +1,33 @@
-import React from 'react';
-import { Provider } from 'react-redux';
-import { store } from '../redux-feature/store';
+import React, { useCallback, useState, createContext } from 'react';
+import { createSessionStorage } from '../helper/createStorage';
 
-function AppContext({ children }) {
-  return <Provider store={store}>{children}</Provider>;
+export const AppContext = createContext();
+
+function AppProvider({ children }) {
+  const activeStorage = createSessionStorage('activeTab');
+
+  const [activeTab, setActiveTab] = useState({
+    tab: activeStorage.get('tab') || 'home',
+    subTab: activeStorage.get('subTab') || ''
+  });
+
+  if (!activeStorage.get('tab')) {
+    // !undefined, !null, !"" -> true
+    activeStorage.set('tab', activeTab.tab);
+    activeStorage.set('subTab', activeTab.subTab);
+  }
+
+  const handleSetActiveTab = useCallback((tab, subTab) => {
+    activeStorage.set('tab', tab);
+    activeStorage.set('subTab', subTab);
+    setActiveTab({ ...activeTab, tab, subTab });
+  }, []);
+
+  return (
+    <AppContext.Provider value={{ activeTab, handleSetActiveTab }}>
+      {children}
+    </AppContext.Provider>
+  );
 }
 
-export default AppContext;
+export default AppProvider;
